+++ packages/jaspr/CHANGELOG.md
@@ -3,6 +3,8 @@
 - Promoted `jaspr_web_compilers` to non-experimental status and changed cli command `jaspr create --experimental-web-compilers` 
   to `jaspr create --jaspr-web-compilers`
 - Fixed error on windows when running `jaspr build`.
+- Added StyleSheet object to jaspr/lib/src/components/style.dart, so you can add custom CSS StyleSheets
+- Modified _BaseDocument to inlcude option StyleSheet list attribute
 
 ## 0.9.3